fix(携程撞库任务):
Integer、Long类型转换优化

+++ marketing-cs/src/main/java/com/br/marketing/service/Impl/PushDataServiceImpl.java
@@ -1274,7 +1274,7 @@ public class PushDataServiceImpl implements PushDataService {
     public void  pushXieChengSmsCollidingToDbData(String data) {
         try {
             JSONObject jsonObject = JSONObject.parseObject(data);
-            Long localId = Long.valueOf(jsonObject.getInteger("localId"));
+            Long localId = jsonObject.getLong("localId");
             Boolean isNewFile = jsonObject.getBooleanValue("isNewFile");
             // 创建线程池
             ThreadPoolExecutor xieChengSmsCollidingThread =
